Moves inner and outer SNARK into Paramters in preparation for Transaction move
