+++ dpc/src/circuits/mod.rs
@@ -22,3 +22,6 @@ pub use inner_circuit_gadget::*;
 
 pub mod inner_circuit_verifier_input;
 pub use inner_circuit_verifier_input::*;
+
+pub mod outer_circuit_verifier_input;
+pub use outer_circuit_verifier_input::*;
