+++ dpc/src/testnet1/outer_circuit/mod.rs
@@ -19,6 +19,3 @@ pub use outer_circuit::*;
 
 pub mod outer_circuit_gadget;
 pub use outer_circuit_gadget::*;
-
-pub mod outer_circuit_verifier_input;
-pub use outer_circuit_verifier_input::*;
